separated dataclass classes from model module, references #16
