+++ src/vcclib/model.py
@@ -1,50 +1,10 @@
 import re
 
-from dataclasses import dataclass
-from dataclasses import field
-from datetime import datetime
 from difflib import SequenceMatcher
 from sqlobject import *
 from sqlobject.sqlbuilder import Select
-from typing import List
 
 from vcclib import database
-
-@dataclass
-class CountedStop:
-    id: int
-    international_id: str = None
-    name: str
-    latitude: float = 0.0
-    longitude: float = 0.0
-    sequence: int
-
-    def __init__(self) -> None:
-        pass
-
-@dataclass
-class CountingSequence:
-    door_id: str
-    area_id: str
-    object_class: str
-    begin_timestamp: datetime
-    end_timestamp: datetime
-    count_in: int = 0
-    count_out: int = 0
-
-    def __init__(self) -> None:
-        pass
-
-@dataclass
-class PassengerCountingEvent:
-    latitude: float = 0.0
-    longitude: float = 0.0
-    after_stop_sequence: int = -1
-    counted_stop: CountedStop = None
-    counting_sequences: List[CountingSequence] = field(default_factory=list)
-
-    def __init__(self) -> None:
-        pass
 
 class Stop(SQLObject):
     stop_id = IntCol()
